// face-check-ui/src/app/services/models/page-response-work-site-response.ts
/* tslint:disable */
/* eslint-disable */
/* Code generated by ng-openapi-gen DO NOT EDIT. */

import { WorkSiteResponse } from '../models/work-site-response';
export interface PageResponseWorkSiteResponse {
  content?: Array<WorkSiteResponse>;
  first?: boolean;
  last?: boolean;
  number?: number;
  size?: number;
  totalElement?: number;
  totalPages?: number;
}